+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file
 - Option to declare multiple usages in one statement: use System, String;
 - Functionality to add any combination of LIST and OBJECT expressions
 - Bulk operations and multiple select on robots and projects in the project view
+- Encode.stringToBase64 construct
+- Decode.stringFromBase64 construct
 
 
 ## [3.0.23] - 11-12-2015